requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.9.3
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asteroid-filterbanks==0.4.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.4
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
colorlog==6.8.0
comm==0.2.0
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
docopt==0.6.2
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.9.0
fonttools==4.47.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
greenlet==3.0.3
httplib2==0.20.2
huggingface-hub==0.20.2
HyperPyYAML==1.2.2
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.26.0
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
julius==0.2.7
jupyter-archive==3.4.0
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
librosa==0.10.1
lightning==2.1.3
lightning-utilities==0.10.0
llvmlite==0.41.1
lxml==4.9.3
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
noisereduce==3.0.0
notebook==6.5.5
notebook_shim==0.2.3
numba==0.58.1
numpy==1.24.1
oauthlib==3.2.0
omegaconf==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
optuna==3.5.0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
Pillow==9.3.0
platformdirs==3.11.0
pooch==1.8.0
primePy==1.3
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.2
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.audio @ git+https://github.com/pyannote/pyannote-audio@8f477fada9a595dc0c0e54de92e8c43cd9b1a7ea
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.21
pydub==0.25.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.1.3
pytorch-metric-learning==2.4.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.12.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.11.4
SecretStorage==3.3.1
semver==3.0.2
Send2Trash==1.8.2
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==0.5.16
SQLAlchemy==2.0.25
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.5.2
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.0+cu118
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.1.0+cu118
torchmetrics==1.2.1
torchvision==0.16.0+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==1.26.13
wadllib==1.3.6
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
yarl==1.9.4
zipp==1.0.0
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart==0.0.6
aiofiles==23.2.1